Added Plugin.cs && change version c# && added description to Config.cs && added to Plugin.cs requirexiledversion
