+++ BetterBlueCandyHalloween/Config.cs
@@ -1,14 +1,14 @@
+using System.ComponentModel;
 using Exiled.API.Interfaces;
 
-namespace BetterBlueCandyHalloween
+namespace BetterBlueCandyHalloween;
+public class Config : IConfig
 {
-    public class Config : IConfig
-    {
-        public bool IsEnabled { get; set; } = true;
-
-        public bool Debug { get; set; }
-
-        public float TimeBlueCandy { get; set; } = 20f;
-        public float ShieldBlueCandy { get; set; } = 450f;
-    }
+    public bool IsEnabled { get; set; } = true;
+    public bool Debug { get; set; }
+    
+    [Description("Time for blue candy")]
+    public float TimeBlueCandy { get; set; } = 20f;
+    [Description("Custom shield to blue candy")]
+    public float ShieldBlueCandy { get; set; } = 450f;
 }
